Gateway/requirements.txt
Flask
flask-sqlalchemy
psycopg2
psycopg2-binary
flask-wtf
git+https://github.com/imwithye/flask-inject.git@master
PyJWT
requests
requests-oauthlib
git+https://github.com/mitsuhiko/flask-oauth.git@master